requirements.txt
streamlit
pandas
scikit-learn
scikit-surprise
flask
